refactor: migrate label
