+++ packages/components/label/src/label.js
