+++ packages/components/label/src/label.visualspec.js
