+++ packages/components/label/src/required-indicator.js
