+++ packages/components/label/src/required-indicator.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../test-utils';
+import { render } from '../../../../src/test-utils';
 import RequiredIndicator from './required-indicator';
 
 it('should render', () => {
