+++ packages/components/label/src/version.js
